% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_match.r
\name{base_match}
\alias{base_match}
\title{\code{base_match()}}
\usage{
base_match(original_variable, codebook)
}
\arguments{
\item{original_variable}{the original variable}

\item{codebook}{the codebook, specified as a named vector in "label=level" format, with the levels listed in the desired order}
}
\value{
a factor
}
\description{
A base-R approximation of dplyr's \code{case_match()}. Unlike \code{case_match()}, \code{base_when()} returns a factor. The levels will be ordered according to the order included in \code{codebook}.
}
\examples{
dmdborn4_codebook<-c('USA'=1,'Other'=2)
ii$country<-base_match(ii$dmdborn4,dmdborn4_codebook)
}
