% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_when.r
\name{base_when}
\alias{base_when}
\title{\code{base_when()}}
\usage{
base_when(conditions)
}
\arguments{
\item{conditions}{conditions for defining the replacement values, specified as a named list, with each element in \code{'label'=logical_vector} format, with the levels listed in the desired order}
}
\value{
a factor
}
\description{
A base-R approximation of dplyr's \code{case_when()}. Unlike \code{case_when()}, \code{base_when()} returns a factor. The levels will be ordered according to the order included in \code{conditions}.
}
\examples{
ii$cholesterol<-base_when(list(
   'Desirable'=ii$lbxtc<200,
   'Borderline high'=(ii$lbxtc>=200)&(ii$lbxtc<240),
   'High'=ii$lbxtc>=240
))
}
